Extracted most of the rest of the dialog box setup code to separate files. Renamed all_ui.py to app_state.py
