--- app/ui/all_ui.py
@@ -1,3 +1,0 @@
-TIMEOUT_DISPLAY_LIST = [["10 sec", 10]]
-for n in range(1, 13):
-    TIMEOUT_DISPLAY_LIST.append([str(n * 10) + " min", n * 600])
